module3/homework/requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
pandas-gbq
